fixed fprintf int size warnings

+++ lib/cUtils/specs/verboseRunner.cpp
@@ -98,8 +98,8 @@ bool VerboseRunner::assertShouldEqual(bool sense,
   bool condition = (actualValue == expectedValue);
   if (condition != sense) {
     fprintf(logFile, "-->>> should %s be equal\n", (sense ? "" : "not"));
-    fprintf(logFile, "---->   actual: [%s] = %ld\n", actualStr, actualValue);
-    fprintf(logFile, "----> expected: [%s] = %ld\n", expectedStr, expectedValue);
+    fprintf(logFile, "---->   actual: [%s] = %lld\n", actualStr, actualValue);
+    fprintf(logFile, "----> expected: [%s] = %lld\n", expectedStr, expectedValue);
     fprintf(logFile, "----> file: %s(%zu)\n", fileName, lineNum);
     SHOULD_FAILED;
   } else numSuccessfulShoulds++;
@@ -189,6 +189,6 @@ void VerboseRunner::logReport(void) {
   fprintf(logFile, "       Specs: %zu \t%zu\t\t%zu\t%zu\n",
           numFailedSpecs, numSuccessfulSpecs, numPendingSpecs, numSpecs);
   fprintf(logFile, "     Shoulds: %zu \t%zu\t\t%zu\t%zu\n",
-          numFailedShoulds, numSuccessfulShoulds, 0L, numShoulds);
+          numFailedShoulds, numSuccessfulShoulds, (size_t)0, numShoulds);
   fprintf(logFile, "------------------------------------------------------\n");
 }
